Mostly Works
The way we declare categories in the model is weird and a bit problematic. That is why it is not yet possible to assign items to collections based on category. The category properties for collections and items are not taken by the db. As soon as we find a way to seed collections/items with categories as they should, making collection with the right items in them will be no problem. The script is already written.